Revert to v1.3.1
Accidently merged dev experiments to master. This commit reverts
to the previous stable version. See "dev" branch for the changes.

+++ rtree_test.go
@@ -132,7 +132,7 @@ func rSaneRect(r rect) error {
 	return nil
 }
 
-func rSaneNode(tr *RTree, r *rect, height int16) error {
+func rSaneNode(tr *RTree, r *rect, height int) error {
 	n := r.data.(*node)
 	if n.count > maxEntries {
 		return errors.New("invalid count")
